// ui/src/features/investigations/api/investigation-delete.generated.ts
type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] }
// ...
export type InvestigationDeleteMutationVariables = Exact<{
  id: string
}>

export type InvestigationDeleteMutation = {
  readonly investigationDelete: boolean
}

export const InvestigationDeleteDocument = {
  kind: "Document",
  definitions: [
    {
      kind: "OperationDefinition",
      operation: "mutation",
      name: { kind: "Name", value: "InvestigationDelete" },
      variableDefinitions: [
        {
          kind: "VariableDefinition",
          variable: { kind: "Variable", name: { kind: "Name", value: "id" } },
          type: {
            kind: "NonNullType",
            type: {
              kind: "NamedType",
              name: { kind: "Name", value: "String" },
            },
          },
        },
      ],
      selectionSet: {
        kind: "SelectionSet",
        selections: [
          {
            kind: "Field",
            name: { kind: "Name", value: "investigationDelete" },
            arguments: [
              {
                kind: "Argument",
                name: { kind: "Name", value: "id" },
                value: {
                  kind: "Variable",
                  name: { kind: "Name", value: "id" },
                },
              },
            ],
          },
        ],
      },
    },
  ],
} as unknown as DocumentNode<InvestigationDeleteMutation, InvestigationDeleteMutationVariables>
// ...
export const InvestigationDeleteMutationSchema: z.ZodType<InvestigationDeleteMutation> = z.object({
  investigationDelete: z.boolean(),
})