// ui/src/features/investigations/api/investigation-detail.generated.ts
type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] }
// ...
export type InvestigationDetailQueryVariables = Exact<{
  id: string
}>

export type InvestigationDetailQuery = {
  readonly investigation: {
    readonly id: string
    readonly name: string
    readonly state: string
    readonly createdAtNanos: string
    readonly updatedAtNanos: string
  } | null
}

export const InvestigationDetailDocument = {
  kind: "Document",
  definitions: [
    {
      kind: "OperationDefinition",
      operation: "query",
      name: { kind: "Name", value: "InvestigationDetail" },
      variableDefinitions: [
        {
          kind: "VariableDefinition",
          variable: { kind: "Variable", name: { kind: "Name", value: "id" } },
          type: {
            kind: "NonNullType",
            type: {
              kind: "NamedType",
              name: { kind: "Name", value: "String" },
            },
          },
        },
      ],
      selectionSet: {
        kind: "SelectionSet",
        selections: [
          {
            kind: "Field",
            name: { kind: "Name", value: "investigation" },
            arguments: [
              {
                kind: "Argument",
                name: { kind: "Name", value: "id" },
                value: {
                  kind: "Variable",
                  name: { kind: "Name", value: "id" },
                },
              },
            ],
            selectionSet: {
              kind: "SelectionSet",
              selections: [
                { kind: "Field", name: { kind: "Name", value: "id" } },
                { kind: "Field", name: { kind: "Name", value: "name" } },
                { kind: "Field", name: { kind: "Name", value: "state" } },
                {
                  kind: "Field",
                  name: { kind: "Name", value: "createdAtNanos" },
                },
                {
                  kind: "Field",
                  name: { kind: "Name", value: "updatedAtNanos" },
                },
              ],
            },
          },
        ],
      },
    },
  ],
} as unknown as DocumentNode<InvestigationDetailQuery, InvestigationDetailQueryVariables>
// ...
export const InvestigationDetailQuerySchema: z.ZodType<InvestigationDetailQuery> = z.object({
  investigation: z
    .object({
      id: z.string(),
      name: z.string(),
      state: z.string(),
      createdAtNanos: z.string(),
      updatedAtNanos: z.string(),
    })
    .nullable(),
})