// ui/src/features/investigations/api/investigations-list.generated.ts
type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] }
// ...
export type InvestigationsListQueryVariables = Exact<{ [key: string]: never }>

export type InvestigationsListQuery = {
  readonly investigations: ReadonlyArray<{
    readonly id: string
    readonly name: string
    readonly state: string
    readonly updatedAtNanos: string
    readonly createdAtNanos: string
  }>
}

export const InvestigationsListDocument = {
  kind: "Document",
  definitions: [
    {
      kind: "OperationDefinition",
      operation: "query",
      name: { kind: "Name", value: "InvestigationsList" },
      selectionSet: {
        kind: "SelectionSet",
        selections: [
          {
            kind: "Field",
            name: { kind: "Name", value: "investigations" },
            selectionSet: {
              kind: "SelectionSet",
              selections: [
                { kind: "Field", name: { kind: "Name", value: "id" } },
                { kind: "Field", name: { kind: "Name", value: "name" } },
                { kind: "Field", name: { kind: "Name", value: "state" } },
                {
                  kind: "Field",
                  name: { kind: "Name", value: "updatedAtNanos" },
                },
                {
                  kind: "Field",
                  name: { kind: "Name", value: "createdAtNanos" },
                },
              ],
            },
          },
        ],
      },
    },
  ],
} as unknown as DocumentNode<InvestigationsListQuery, InvestigationsListQueryVariables>
// ...
export const InvestigationsListQuerySchema: z.ZodType<InvestigationsListQuery> = z.object({
  investigations: z.array(
    z.object({
      id: z.string(),
      name: z.string(),
      state: z.string(),
      updatedAtNanos: z.string(),
      createdAtNanos: z.string(),
    })
  ),
})